Move search index entry types to dedicated files
